refactor, uso del service
