+++ frontend/src/services/ClienteService.ts
@@ -9,13 +9,24 @@ class ClienteService {
   constructor(path: string) {
     this.path = path;
   }
-
+  getClienteById(clienteId: string): Promise<AxiosResponse<Cliente>> {
+    return HttpClient.get<Cliente>(`${this.path}/${clienteId}`);
+  }
   getAllClientes(): Promise<AxiosResponse<Cliente[]>> {
     return HttpClient.get<Cliente[]>(`${this.path}/misClientes`);
   }
 
   getAllCamposByClienteId(clienteId: string): Promise<AxiosResponse<Campo[]>> {
     return HttpClient.get<Campo[]>(`${this.path}/${clienteId}/campos`);
+  }
+  createCliente(nuevoCliente: Partial<Cliente>): Promise<AxiosResponse<Cliente>> {
+    return HttpClient.post<Cliente>(`${this.path}`, nuevoCliente);
+  }
+  updateCliente(clienteId: number, updatedCliente: Cliente): Promise<AxiosResponse<Cliente>> {
+    return HttpClient.put<Cliente>(`${this.path}/${clienteId}`, updatedCliente);
+  }
+  deleteCliente(clienteId: number): Promise<AxiosResponse<void>> {
+    return HttpClient.delete<void>(`${this.path}/${clienteId}`);
   }
 }
